refactor(models): relocate Documentation BlockElements/MsrQuery modules out of TextModel (Rule 0007 A)
