+++ src/armodel/models/M2/MSR/Documentation/BlockElements/Figure.py
@@ -1,5 +1,5 @@
 from armodel.models.M2.MSR.Documentation.TextModel.LanguageDataModel import LanguageSpecific
-from armodel.models.M2.MSR.Documentation.TextModel.BlockElements.PaginationAndView import Paginateable
+from armodel.models.M2.MSR.Documentation.BlockElements.PaginationAndView import Paginateable
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.ArObject import ARObject
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.EngineeringObject import EngineeringObject
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import AREnum
