+++ tests/test_armodel/models/M2/MSR/Documentation/BlockElements/test_PaginationAndView.py
@@ -1,10 +1,10 @@
 """
-This module contains tests for the PaginationAndView module in MSR.Documentation.TextModel.BlockElements.
+This module contains tests for the PaginationAndView module in MSR.Documentation.BlockElements.
 """
 
 import pytest
 
-from armodel.models.M2.MSR.Documentation.TextModel.BlockElements.PaginationAndView import (
+from armodel.models.M2.MSR.Documentation.BlockElements.PaginationAndView import (
     DocumentViewSelectable,
     Paginateable,
 )
