+++ tests/test_armodel/models/M2/MSR/Documentation/BlockElements/test_RequirementsTracing.py
@@ -1,8 +1,8 @@
 """This module contains tests for the RequirementsTracing module."""
 
 from armodel.models.M2.AUTOSARTemplates.GenericStructure.GeneralTemplateClasses.PrimitiveTypes import RefType, String
+from armodel.models.M2.MSR.Documentation.BlockElements.RequirementsTracing import StructuredReq, TraceableText
 from armodel.models.M2.MSR.Documentation.TextModel.BlockElements import DocumentationBlock
-from armodel.models.M2.MSR.Documentation.TextModel.BlockElements.RequirementsTracing import StructuredReq, TraceableText
 
 
 class TestTraceableText:
