+++ tests/test_armodel/models/M2/MSR/Documentation/test_Chapters.py
@@ -11,15 +11,13 @@ from armodel.models.M2.MSR.Documentation.Chapters import (
     ChapterContent,
     ChapterModel,
     ChapterOrMsrQuery,
-    MsrQueryChapter,
-    MsrQueryTopic1,
     PredefinedChapter,
     Topic1,
     TopicContent,
     TopicContentOrMsrQuery,
     TopicOrMsrQuery,
 )
-from armodel.models.M2.MSR.Documentation.TextModel.MsrQuery import MsrQueryP1
+from armodel.models.M2.MSR.Documentation.MsrQuery import MsrQueryChapter, MsrQueryP1, MsrQueryTopic1
 
 
 class TestChapter:
